Fix embed pdf update. Disposition type in dev server.
The embedded pdf reader did not update when there are
multiple files in some production scenarios. To fix this
a key was added to the <embed> component in ImageCollection to
force react to rerender and thereby update component.
Also changed disposition type of files served from dev server to
inline from previously deprecated "attachment".