Add feedback field to ExerciseMeta
Successfully added feedback as a meta.  The lack of correct field has to
be compensated in questiontypes.  Took away feedback option in xml and
added code to respect feedback in courselist and exerciselist.  It is
now backward compatible with main branch, with diverging behavior only
if nofeedback is checked.  Warning is issued if nofeedback is tried on
any questiontype except devLinearAlgebra.