Add tag QMath and AsciiMath
QMath renders mathematics in the context of a question, using the
variables and other data associated with that question type. This is
based on the work in mathtag6 but with considerable changes in methods:
The QMath component is separate from Question but instanciates the
corresponding question type class and uses its renderAsciiMath method.
This enabled arbitrary nesting (not possible with Question tag) and
decouples the heavy processing associated with a full question render.

AsciiMath tag renders ascii math using mathjs with the same
preprocessing used by the various question types but without special
handling of defined variables etc.