Fix SUBPATH and MEDIA_URL
MEDIA_URL now follows the official django docs with the
format "/media/".
Updated urls to reflext this change.
Renamed serve_media to serve_media_public and fixed the paths
according to the above changes.

Added frontend settings.js as tracked file with cookie handling
for subpath since this falls back to empty if not found.

Update serve_file to be compatible with not using dev_path for
standard sittuations but still getting the correct path for
local development.