Add OpenTAUser model
New model with one-to-one field to django User model. Used to
extend a user with more data than what the User model provides.
Here a new field courses is added that indicates which courses
a user is registered for.

Update registration by domain to handle course provided in url.
This is propagated from login site that has been called with
https://site_url/course_name

If a user already exists, the course is just added to the list
above and the user is redirected to the login screen. Otherwise
the registration works as usual.

TODO: Fix course spinner on course change
TODO: Fix registration by password