Refactor text rendering in questions
Moved out text rendering functionality to joint module. This can
optionally be used by a question type to support translations.
The function

  renderText(json, dispatcher, language)

will render it's $ content (and children using dispatcher) if no alts
are present. If there are children alt with the selected language this
will be rendered instead.

If no dispatcher is provided (null), a default dispatcher is used that
only renders text.

The possibility to pass on the dispatcher is useful when the text tag
can contain more advanced children (such as in multiple choice where we
allow figures in choice texts).