Add canvas gradebook component
Add gradebook component that enables a teacher
to upload a gradebook generated from canvas with

	Grades->Actions->Export

The uploaded CSV file is the processed in the following way

All fields except the required fields according to
https://community.canvaslms.com/docs/DOC-17747-41521142911
are removed.

New fields corresponding to required and bonus exercises with&without
deadline are added.

The resulting file is returned to the user that can then be uploaded to
Canvas to update the results with

	Grades->Actions->Import

If this is done twice the corresponding fields will be updated.